rename serial_ros to serial_node since everything here should be a ros package
