+++ src/serial_node/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/serial_node
+[install]
+install_scripts=$base/lib/serial_node
