+++ src/serial_node/setup.py
@@ -1,6 +1,6 @@
 from setuptools import find_packages, setup
 
-package_name = 'serial_ros'
+package_name = 'serial_node'
 
 setup(
     name=package_name,
@@ -20,7 +20,7 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'op_reader = serial_ros.node:main'
+            'op_reader = serial_node.node:main'
         ],
     },
 )
