+++ src/serial_node/test/test_copyright.py
