--- src/serial_ros/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/serial_ros
-[install]
-install_scripts=$base/lib/serial_ros
